refactor photo code
